Initial refactoring work to disentangle PS specifics
