+++ src/Microsoft.Management.Configuration.Processor/PowerShell/DscModules/IDscModule.cs
@@ -1,14 +1,14 @@
-﻿// -----------------------------------------------------------------------------
+// -----------------------------------------------------------------------------
 // <copyright file="IDscModule.cs" company="Microsoft Corporation">
 //     Copyright (c) Microsoft Corporation. Licensed under the MIT License.
 // </copyright>
 // -----------------------------------------------------------------------------
 
-namespace Microsoft.Management.Configuration.Processor.DscModule
+namespace Microsoft.Management.Configuration.Processor.PowerShell.DscModules
 {
     using System.Collections.Generic;
     using System.Management.Automation;
-    using Microsoft.Management.Configuration.Processor.DscResourcesInfo;
+    using Microsoft.Management.Configuration.Processor.PowerShell.DscResourcesInfo;
     using Microsoft.PowerShell.Commands;
     using Windows.Foundation.Collections;
