+++ src/Microsoft.Management.Configuration.Processor/PowerShell/Extensions/PowerShellExtensions.cs
