+++ src/Microsoft.Management.Configuration.Processor/PowerShell/ProcessorEnvironments/HostedEnvironment.cs
@@ -15,15 +15,16 @@ namespace Microsoft.Management.Configuration.Processor.Runspaces
     using System.Management.Automation.Runspaces;
     using System.Runtime.InteropServices.WindowsRuntime;
     using Microsoft.Management.Configuration.Processor.Constants;
-    using Microsoft.Management.Configuration.Processor.DscModule;
-    using Microsoft.Management.Configuration.Processor.DscResourcesInfo;
     using Microsoft.Management.Configuration.Processor.Helpers;
+    using Microsoft.Management.Configuration.Processor.PowerShell.DscModules;
+    using Microsoft.Management.Configuration.Processor.PowerShell.DscResourcesInfo;
+    using Microsoft.Management.Configuration.Processor.PowerShell.Helpers;
     using Microsoft.Management.Configuration.Processor.ProcessorEnvironments;
     using Microsoft.PowerShell.Commands;
     using Windows.Foundation.Collections;
     using Windows.Security.Cryptography.Certificates;
     using Windows.Storage.Streams;
-    using static Microsoft.Management.Configuration.Processor.Constants.PowerShellConstants;
+    using static Microsoft.Management.Configuration.Processor.PowerShell.Constants.PowerShellConstants;
 
     /// <summary>
     /// Process environment. Provides interaction with PowerShell for a hosted environment.
