+++ src/Microsoft.Management.Configuration.UnitTests/Tests/DscModuleV2SimpleFileResourceTests.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Management.Configuration.UnitTests.Tests
     using System.IO;
     using System.Management.Automation;
     using Microsoft.Management.Configuration.Processor.DscModule;
-    using Microsoft.Management.Configuration.Processor.Helpers;
+    using Microsoft.Management.Configuration.Processor.PowerShell.Helpers;
     using Microsoft.Management.Configuration.UnitTests.Fixtures;
     using Microsoft.Management.Configuration.UnitTests.Helpers;
     using Windows.Foundation.Collections;
